# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: tts
    description: Text to speech operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /tts/voices:
    get:
      operationId: getVoices
      tags: [tts]
      summary: Get available voices
      description: Returns list of available ElevenLabs voices
      responses:
        "200":
          description: List of voices
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoicesResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tts/generate:
    post:
      operationId: generateSpeech
      tags: [tts]
      summary: Generate speech from text
      description: Converts text to speech using ElevenLabs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateSpeechRequest"
      responses:
        "200":
          description: Audio file
          content:
            audio/mpeg:
              schema:
                type: string
                format: binary
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Voice:
      type: object
      properties:
        voice_id:
          type: string
        name:
          type: string
        category:
          type: string
      required:
        - voice_id
        - name
        - category
    VoicesResponse:
      type: object
      properties:
        voices:
          type: array
          items:
            $ref: "#/components/schemas/Voice"
      required:
        - voices
    GenerateSpeechRequest:
      type: object
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 5000
        voice_id:
          type: string
        model_id:
          type: string
      required:
        - text
        - voice_id
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error