# lib/api-zod/src/generated/types/generateSpeechRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateSpeechRequest {
  /**
   * @minLength 1
   * @maxLength 5000
   */
  text: string;
  voice_id: string;
  model_id?: string;
}